Got a bit faster python, graphs with memcpy tasks to much time to launch...

`1124  303.19 ms run,   96.72 ms python,  11.78 ms fetch data,  194.70 ms CUDA * 6,  6.06 loss, 0.02 acc, 0.283836 LR, 93.78 GB used,  51043.88 GFLOPS`