@geohot 2024-03-26
"graphs with memcpy tasks to much time to launch" how is gpt-fast so fast?