@geohot 2024-03-26
gpt-fast just works on the computer, you can download it and test it. Was getting quite high llama sharded on 4 GPUs tok/s.